Style 11, key updates

+++ core/src/main/java/com/zarbosoft/merman/core/syntax/alignments/RelativeAlignmentSpec.java
@@ -5,15 +5,15 @@ import com.zarbosoft.merman.core.visual.alignment.RelativeAlignment;
 
 public class RelativeAlignmentSpec implements AlignmentSpec {
   public final String base;
-  public final int offset;
+  public final double offset;
   public final boolean collapse;
 
   public static class Config {
     public final String base;
-    public final int offset;
+    public final double offset;
     public final boolean collapse;
 
-    public Config(String base, int offset, boolean collapse) {
+    public Config(String base, double offset, boolean collapse) {
       this.base = base;
       this.offset = offset;
       this.collapse = collapse;
